requirements.txt
aiohttp==3.5.4
async-timeout==3.0.1
attrs==19.1.0
chardet==3.0.4
discord==1.0.1
EasyProcess==0.2.7
idna==2.8
multidict==4.5.2
Pillow==6.1.0
pyscreenshot==0.5.1
python-dotenv==0.10.3
websockets==6.0
yarl==1.3.0
azure-cognitiveservices-vision-computervision==0.4.0
msrest==0.6.9
